% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tg_lasso.R
\name{backward.ranking}
\alias{backward.ranking}
\title{Backward Ranking}
\usage{
backward.ranking(X, Y, ranked.features, lambda.1.range, lambda.2.range,
  lambda.3.range, k, allow.missing.X = FALSE, allow.missing.Y = FALSE,
  allow.zero.sd = FALSE, intervals = NULL)
}
\arguments{
\item{X}{A feature matrix, with one row per patient and one column per
feature.}

\item{Y}{A longitudinal outcome matrix, with one row per patient and one
column per timepoint.}

\item{lambda.1.range}{A vector of lambda.1 values to try.}

\item{lambda.2.range}{A vector of lambda.2 values to try.}

\item{lambda.3.range}{A vector of lambda.3 values to try.}

\item{allow.missing.X}{If TRUE, missing feature values are imputed using
the average value of the missing feature. Default FALSE.}

\item{allow.missing.Y}{If TRUE, missing outcome values are allowed in the
Y longitudinal outcome matrix. In this case, missing Y values result in zero
errors, and therefore have no influence in the model. Default FALSE.}

\item{allow.zero.sd}{If TRUE, features with zero standard deviations are
allowed in the X matrix. Default FALSE.}

\item{intevals}{An optional vector of time intervals between the time points
in the columns of Y. By default, they are all set to 1.}
}
\value{
A list having the following members:
\itemize{
  \item{"X"}{The original X feature matrix.}
  \item{"Y"}{The original Y longitudinal outcome matrix.}
  \item{"ranked.features"}{The set of features to rank (may be all
    features, for example).}
  \item{"min.index"}{The column index of the feature having the smallest
    mean CV RMSE.}
  \item{"min.feature"}{The feature having the smallest mean CV RMSE.}
  \item{"min.mean"}{The mean CV RMSE associated to the feature having the
    smallest mean CV RMSE.}
  \item{"min.sd"}{The sd of the CV RMSE associated to the feature having the
    smallest mean CV RMSE.}
  \item{"max.index"}{The column index of the feature having the largest
    mean CV RMSE.}
  \item{"max.feature"}{The feature having the largest mean CV RMSE.}
  \item{"max.mean"}{The mean CV RMSE associated to the feature having the
    largest mean CV RMSE.}
  \item{"max.sd"}{The sd of the CV RMSE associated to the feature having the
    largest mean CV RMSE.}
  \item{"results"}{Table of results of the backward ranking."}
}
}
\description{
Eliminates and replaces features one at a time, ranking features based on
the performance obtained by eliminating them from the feature set. The
minimum and maximum values of the mean CV RMSE, the features corresponding to
these values, as well as the mean and standard deviation of the CV RMSE.
}

