% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tg_lasso.R
\name{backward.selection}
\alias{backward.selection}
\title{Backward Selection}
\usage{
backward.selection(X, Y, lambda.1.range, lambda.2.range, lambda.3.range, k,
  allow.missing.X = FALSE, allow.missing.Y = FALSE, allow.zero.sd = FALSE,
  intervals = NULL)
}
\description{
Backward Selection
}

