% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tg_lasso.R
\name{predict.tg.lasso}
\alias{predict.tg.lasso}
\title{Predict Smooth Temporal Model}
\usage{
\method{predict}{tg.lasso}(model, X, lambda.1 = NULL, lambda.2 = NULL,
  lambda.3 = NULL)
}
\description{
Predict Smooth Temporal Model
}

