% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tg_lasso.R
\name{tg.lasso}
\alias{tg.lasso}
\title{Temporal Group LASSO}
\usage{
tg.lasso(X, Y, lambda.1.range, lambda.2.range, lambda.3.range,
  allow.missing.X = FALSE, allow.missing.Y = FALSE, allow.zero.sd = FALSE,
  intervals = NULL, min.difference = 1e-07)
}
\arguments{
\item{X}{A feature matrix, with one row per patient and one column per
feature.}

\item{Y}{A longitudinal outcome matrix, with one row per patient and one
column per timepoint.}

\item{lambda.3.range}{Tuning parameter range for the group penalty for each
feature across time points. It is encouraged to pass multiple values at a
time in order to take advantage of warm starts.}

\item{allow.missing.X}{If TRUE, missing feature values are imputed using
the average value of the missing feature. Default FALSE.}

\item{allow.missing.Y}{If TRUE, missing outcome values are allowed in the
Y longitudinal outcome matrix. In this case, missing Y values result in zero
errors, and therefore have no influence in the model. Default FALSE.}

\item{allow.zero.sd}{If TRUE, features with zero standard deviations are
allowed in the X matrix. Default FALSE.}

\item{lambda.1}{Tuning parameter for the coefficient shrinkage inducing
penalty.}

\item{lambda.2}{Tuning parameter for the temporal smoothness inducing
penalty.}

\item{intevals}{An optional vector of time intervals between the time points
in the columns of Y. By default, they are all set to 1.}
}
\value{
A list having the following members:
\itemize{
  \item{"X"}{The original X feature matrix.}
  \item{"X.mean"}{The vector of column means of X.}
  \item{"X.sd"}{The vector of column standard deviations of X.}
  \item{"Y"}{The original Y longitudinal outcome matrix.}
  \item{"Y.mean"}{The vector of column means of Y.}
  \item{"Y.sd"}{The vector of column standard deviations of Y.}
  \item{"W"}{The matrix of fitted regression coefficients.}
  \item{"lambda.1"}{The value of the lambda.1 used to fit the model.}
  \item{"lambda.2"}{The value of the lambda.2 used to fit the model.}
  \item{"lambda.3"}{The value of the lambda.3 used to fit the model.}
  \item{"allow.missing.X"}{Whether missing X values were allowed when fitting
    the model.}
  \item{"allow.missing.Y"}{Whether missing Y values were allowed when fitting
    the model.}
  \item{"intervals"}{The vector of time intervals between the time points in
  Y.}
}
}
\description{
This function fits a regularized regression model with the following
objective function:
}
\details{
L(W) = arg min _W 0.5 (||Y - XW||_F + 0.5 lambda_1 ||W||_F +
  0.5 lambda_2 ||WR||_F + lambda_3 ||W||_2,1),

where X is a feature matrix, Y is a matrix with one row for each patient and
one column for each timepoint and R is a matrix where R_i,j is 1 if i = j,
-1 if i = j + 1, and 0 otherwise. This is a regularized regression with two
L2 penalties, the first one induces shrinkage of the coefficients. The second
one induces temporal smoothness by penalizing differences between adjacent
coefficients corresponding to the same variable but different timepoints.
}
\examples{
# generate data
n <- 1000
p <- 30
t <- 17
X <- matrix(rnorm(n * p, mean = 3, sd = 1.5), nrow = n, ncol = p)
W <- matrix(rnorm(p * t, mean = 0, sd = 1.2), nrow = p, ncol = t)
Y <- X \%*\% W

# test parameter estimation
X <- matrix(rnorm(n * p, mean = 3, sd = 1.5), nrow = n, ncol = p)
W <- matrix(rnorm((p + 1) * t, mean = 0, sd = 1.2), nrow = (p + 1), ncol = t)
Y <- cbind(rep(1, nrow(X)), X) \%*\% W
model <- tg.lasso(X, Y, 0.1, 0.1, 0.00001)
correlation <- cor(c(W), c(model$W.list[[1]][[1]][[1]]))
cat(sprintf("Parameter correlation: \%s.\\n", formatC(correlation,
  digits = 3)))
stopifnot(correlation > 0.99)
}

